+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Handle posts without thumbnail properly
 - Prevent horizontal scrolling on search results
+- `Watch on FP` button not updating when video changes
 
 ### Changed
 - Do not wait for channels sync to finish before returning channel list